+++ src/vic3_world/characters/vic3_character.h
@@ -27,7 +27,7 @@ struct CharacterOptions
    std::set<std::string> traits;
    std::string origin_tag;
    std::optional<int> origin_country_id;
-   std::optional<int> formation_id;
+   std::optional<int64_t> formation_id;
 };
 
 class Character
@@ -91,7 +91,7 @@ class Character
    std::set<std::string> traits_;
    std::string origin_tag_;  // Where an agitator was exiled from, resolve to ID before HoI
    std::optional<int> origin_country_id_;
-   std::optional<int> formation_id_;  // Commanders with formations are actively hired to lead boats/troops
+   std::optional<int64_t> formation_id_;  // Commanders with formations are actively hired to lead boats/troops
 };
 }  // namespace vic3
